refactor: Rename SectionEditDialog => ModeEditDialog
