+++ src/components/EditDialogs/ModeEditDialog/index.js
@@ -1,0 +1,3 @@
+import ModeEditDialog from './ModeEditDialog'
+
+export default ModeEditDialog
