--- src/components/EditDialogs/SectionEditDialog/index.js
@@ -1,3 +1,0 @@
-import SectionEditDialog from './SectionEditDialog'
-
-export default SectionEditDialog
