+++ src/components/TripItem.jsx
@@ -19,7 +19,7 @@ import {
   getEndPlaceDisplayName,
   getFormattedDuration,
   getModesSortedByDistance,
-  formatTripDistance,
+  getFormattedTripDistance,
   getStartDate,
   computeAndFormatCO2Trip,
   computeAndFormatCO2TripByMode
@@ -57,7 +57,7 @@ export const TripItem = ({ geojson, trip, hasDateHeader }) => {
   const endPlace = useMemo(() => getEndPlaceDisplayName(trip), [trip])
   const duration = useMemo(() => getFormattedDuration(trip), [trip])
   const modes = useMemo(() => getModesSortedByDistance(trip), [trip])
-  const distance = useMemo(() => formatTripDistance(trip), [trip])
+  const distance = useMemo(() => getFormattedTripDistance(trip), [trip])
   const day = useMemo(() => f(getStartDate(trip), 'dddd DD MMMM'), [f, trip])
 
   const formattedCO2 = useMemo(() => {
